Refactor test_rs_basics.py into 4 files
